--- README.md
@@ -1,87 +1,0 @@
-goit python group project - personal assistant "Fellow"
-
-Инструкция на русском
-
-«Приятель»
-«Приятель» - это ваш персональный друг и ассистент с книгой контактов, заметками и некоторыми дополнительными функциями.
-
-«Приятель» умеет:
-сохранять, удалять и изменять контакты в вашей книге контактов
-проверять дни рождения ваших друзей, чем заданное количество дней
-сохранять, удалять и изменять ваши заметки
-упорядочивать ваше древо папок
-
-Установка
-ВАЖНО!
-Для установки данного персонального ассистента вам нужно установить пайтон3 и pip.
-
-- скопируйте репозиторий на ваш компьютер
-- поменяйте директорию на my_fellow
-- type pip install -e .
-- для запуска данного бота вам нужно прописать вот эту команду от имени администратора: python3 my_fellow
-
-«Приятель» работает со следующими командами:
-add
-add record – добавление новой записи
-add phone – добавление нового телефона
-add note – добавление новой заметки
-change
-change record – изменение записи
-change note – изменение заметки
-delete
-delete note – удаление записи
-delete record – удаление записи
-get
-get birthdays – получение количества дней до дня рождения
-find
-find record – поиск записи
-find phone – поиск телефона
-find note – поиск заметки
-sort
-sort folder – сортировка содержимого папки
-show
-show records – показать записи
-show notes – показать заметки
-
-Instruction in English
-«Fellow»
-
-«Fellow» is your personal friend and assistant with a contact book, notes and some additional features.
-
-«Fellow» can:
-• save, delete and edit contacts in your contact book
-• check your friends' birthdays than a given number of days
-• save, delete and modify your notes
-• organize your folder tree
-
-Installation
-IMPORTANT!
-To install this personal assistant you need to install python3 and pip
-
-• copy the repository to your computer
-• change directory to my_fellow
-• type pip install -e .
-• to run this bot you need to register this command as administrator: python3 my_fellow
-
-«Fellow» works with the following commands:
-add
-add record - adding a new record
-add phone - adding a new phone
-add note - add a new note
-change
-change record
-change note - note change
-delete
-delete note - delete a note
-delete record - delete a record
-get
-get birthdays - get the number of days until the birthday
-find
-find record - search for a record
-find phone - phone search
-find note - find a note
-sort
-sort folder - sort the contents of a folder
-show
-show records - show records
-show notes - show notes
